<Issues> #6 - Fixed wrong column index in find all reference
The original solution to get the column index is to use string.indexOf,
but there are some cases that string.indexOf cannot handle correctly.

For example, if we want to find "FS_IOC32_SETFLAGS" in
"\#define F2FS_IOC32_SETFLAGS FS_IOC32_SETFLAGS" by indexOf, we will get
the substring of F2"FS_IOC32_SETFLAGS". This is not we want.

This patch fixes this issue by searching the following regex:
"([a-zA-Z0-9_]*)" + symbol + "([a-zA-Z0-9_]*)"

If we find a string that doesn't matches any group, the string is what
we want. If we find a string that matches anyone of the two groups, this
string should be a substring of another symbol, and we know that is not
we want.

Signed-off-by: Lin Chieh <jaycetyle@gmail.com>